requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
click==8.1.3
contextvars==2.4
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
h11==0.14.0
htmltools==0.1.2
idna==3.4
immutables==0.19
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.1.0
matplotlib==3.6.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
numpy==1.23.4
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.4
scipy==1.9.2
shiny==0.2.7
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tabulate==0.9.0
tools==0.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
uvicorn==0.18.3
websockets==10.3
